Added limits on amount returned by APIs

+++ installPreCommitHook.sh
@@ -1,7 +1,13 @@
 #!/bin/bash
 
-cd .git/hooks/
-ln -sf ../../preCommitHook.sh pre-commit
+if [ ! -e .git/hooks/pre-commit ]
+then
+ cd .git/hooks/
+ ln -sf ../../preCommitHook.sh pre-commit
+ echo Pre commit hook installed
+else
+ echo Pre commit hook already in place
+fi
 
 exit 0
